# Conflicts:
#	cmd/node/config/config.toml
#	epochStart/bootstrap/process.go
#	integrationTests/state/stateTrieSync/stateTrieSync_test.go
#	state/errors.go
#	state/syncer/baseAccountsSyncer.go
#	state/syncer/userAccountsSyncer.go
#	update/factory/exportHandlerFactory.go
#	update/genesis/common.go
#	update/genesis/export.go
#	update/genesis/export_test.go
#	update/interface.go
#	update/mock/accountsDBSyncerStub.go
#	update/mock/epochStartTriesSyncHandlerMock.go
#	update/mock/syncStateStub.go
#	update/sync/coordinator.go
#	update/sync/syncAccountsDBs.go
#	update/sync/syncAccountsDBs_test.go

+++ storage/txcache/score.go
@@ -57,7 +57,7 @@ func (computer *defaultScoreComputer) computeRawScore(params senderScoreParams) 
 	countPow2 := params.count * params.count
 	countScore := math.Log(float64(countPow2)+1) + 1
 
-	rawScore := float64(ppuScoreAdjusted) / countScore
+	rawScore := ppuScoreAdjusted / countScore
 	// We apply the logistic function,
 	// and then subtract 0.5, since we only deal with positive scores,
 	// and then we multiply by 2, to have full [0..1] range.
